+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Stream support in wallet serialization
+
 ### Changed
 - Refactored Wallet class
+- Moved json serialization to ExtensionMethods class
 
 ### Fixed
 - PublicKeyPairGenerator